Move ConsoleInput to AceConsoleInput
